+++ pyccel/codegen/printing/fcode.py
@@ -1195,12 +1195,17 @@ class FCodePrinter(CodePrinter):
         return stmt
 
     def _print_NumpyArray(self, expr):
-        expr_args = (expr.arg,) if isinstance(expr.arg, Variable) else expr.arg
         order = expr.order
+
+        try :
+            cast_func = DtypePrecisionToCastFunction[expr.dtype]
+        except KeyError:
+            errors.report(PYCCEL_RESTRICTION_TODO, severity='fatal')
+        arg = expr.arg if expr.arg.dtype == expr.dtype else cast_func(expr.arg)
         # If Numpy array is stored with column-major ordering, transpose values
         # use reshape with order for rank > 2
         if expr.rank <= 2:
-            rhs_code = self._print(expr.arg)
+            rhs_code = self._print(arg)
             if expr.arg.order and expr.arg.order != expr.order:
                 rhs_code = f'transpose({rhs_code})'
             if expr.arg.rank < expr.rank:
@@ -1210,6 +1215,8 @@ class FCodePrinter(CodePrinter):
                     shape_code = ', '.join(self._print(i) for i in expr.shape[::-1])
                 rhs_code = f"reshape({rhs_code}, [{shape_code}])"
         else:
+            expr_args = (expr.arg,) if isinstance(expr.arg, Variable) else expr.arg
+            expr_args = tuple(a if a.dtype == expr.dtype else cast_func(a) for a in expr_args)
             new_args = []
             inv_order = 'C' if order == 'F' else 'F'
             for a in expr_args: